refactor: move allValues, mapValues to objectTools
 - preserve key type in mapValues()
